pointerevents/pointerevent_touch-action-pan-x-pan-y-symmetry_touch.html

<!doctype html>
<html>
    <head>
        <title>touch-action: pan-x pan-y</title>
        <meta name="viewport" content="width=device-width">
        <link rel="stylesheet" type="text/css" href="pointerevent_styles.css">
        <script src="http://w3c-test.org/resources/testharness.js"></script>
        <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
        <script src="pointerevent_support.js"></script>
        <style>
            #target0 {
            width: 40%;
            height: 430px;
            touch-action: pan-x pan-y;
            }
            #target1 {
            background: black;
            color: white;
            white-space: nowrap;
            overflow-y: auto;
            overflow-x: auto;
            width: 40%;
            height: 430px;
            float: right;
            touch-action: pan-y pan-x;
            }
        </style>
    </head>
    <body onload="run()">
        <h2>Pointer Events touch-action attribute support</h2>
        <h4 id="desc">Test Description: Try to scroll text in the <i>left</i> div DOWN. Wait for description update. Expected: pan enabled</h4>
        <p>Note: this test is for touch-devices only</p>
        <div>
            <div id="target1">
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
            </div>
            <div id="target0">
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>
                    Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diem
                    nonummy nibh euismod tincidunt ut lacreet dolore magna aliguam erat volutpat.
                    Ut wisis enim ad minim veniam, quis nostrud exerci tution ullamcorper suscipit
                    lobortis nisl ut aliquip ex ea commodo consequat.
                </p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
                <p>Lorem ipsum dolor sit amet...</p>
            </div>
        </div>
        <script type='text/javascript'>
            var detected_pointertypes = {};

            var xScrollIsReceived = false;
            var yScrollIsReceived = false;
            var xScr0, yScr0, xScr1, yScr1;

            var isSecondPart = false;

            add_completion_callback(showPointerTypes);

            function run() {
                var target0 = document.getElementById("target0");
                var target1 = document.getElementById("target1");

                var test_touchaction_left = async_test("touch-action attribute test: pan-x pan-y");
                var test_touchaction_right = async_test("touch-action attribute test: pan-y pan-x");

                xScr0 = target0.scrollLeft;
                yScr0 = target0.scrollTop;

                on_event(target0, 'pointercancel', function(event) {
                    detected_pointertypes[event.pointerType] = true;
                });

                // left div scroll
                on_event(target0, 'scroll', function(event) {
                    if(!isSecondPart) {
                        xScr1 = target0.scrollLeft;
                        yScr1 = target0.scrollTop;

                        if(xScr1 != xScr0) {
                            xScrollIsReceived = true;
                        }

                        if(yScr1 != yScr0) {
                            test_touchaction_left.step(function () {
                                yScrollIsReceived = true;
                                assert_true(true, "y-scroll received.");
                            });
                            updateDescriptionSecondStepSymmetry();
                        }

                        if(xScrollIsReceived && yScrollIsReceived) {
                            test_touchaction_left.done();
                            isSecondPart = true;
                            xScrollIsReceived = false;
                            yScrollIsReceived = false;
                            xScr0 = target1.scrollLeft;
                            yScr0 = target1.scrollTop;
                            updateDescriptionThirdStepSymmetry();
                        }
                    }
                });

                // right div scroll
                on_event(target1, 'scroll', function(event) {
                    if(isSecondPart) {
                        xScr1 = target1.scrollLeft;
                        yScr1 = target1.scrollTop;

                        if(xScr1 != xScr0) {
                            xScrollIsReceived = true;
                        }

                        if(yScr1 != yScr0) {
                            test_touchaction_right.step(function () {
                                yScrollIsReceived = true;
                                assert_true(true, "y-scroll received.");
                            });
                            updateDescriptionFourthStepSymmetry();
                        }

                        if(xScrollIsReceived && yScrollIsReceived) {
                            test_touchaction_right.done();
                            updateDescriptionComplete();
                        }
                    }
                });
            }

            function updateDescriptionSecondStepSymmetry() {
                document.getElementById('desc').innerHTML = "Test Description: Try to scroll text in the <i>left</i> div RIGHT.";
            }

            function updateDescriptionThirdStepSymmetry() {
                document.getElementById('desc').innerHTML = "Test Description: Try to scroll text in the <i>right</i> div DOWN.";
            }

            function updateDescriptionFourthStepSymmetry() {
                document.getElementById('desc').innerHTML = "Test Description: Try to scroll text in the <i>right</i> div RIGHT.";
            }
        </script>
        <h1>touch-action: pan-x pan-y and pan-y pan-x</h1>
        <div id="complete-notice">
            <p>The following pointer types were detected: <span id="pointertype-log"></span>.</p>
        </div>
        <div id="log"></div>
    </body>
</html>
